Refactor test cases for diagnostic reporting
