+++ projects/testdata/project-with-bad-bala-dep/lib.bal
@@ -14,8 +14,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import testorg/cyclic.moduleA;
+import mockorg/errorpkg as _;
 
-public function fromB() returns string {
-    return moduleA:fromA();
+public function hello() {
 }
