--- projects/testdata/project-with-bad-bala-dep/main.bal
@@ -1,4 +1,0 @@
-import mockorg/errorpkg as _;
-
-public function main() {
-}
